fix: restore build and tests after the dependencies refactor (#265)
* chore: refactore dependencies management

* fix(hosting): restore DI graph and CLI behavior broken by the dependencies refactor

- RegisterConfigCommands, extraction, workspace, and file-ingestion services the
  composition roots call but never registered (ConfigCommands in AppRunner graph;
  RegisterExtractionServices/RegisterWorkspaceService were dead code; ingestion moved
  into core since SqliteMemoryStore now requires IFileIngestor)
- Register concretes the tests resolve: SqliteConnectionFactory, ToolCallMetrics
- Fix inverted success/failure ternary in CliOptionReader.ReadOptions and the
  TryParse failure-path return, restoring parse-error refusal
- Restore serve root-port fallback and read bool flags with requireTokens:false
  (System.CommandLine bool options carry no tokens)
- Restore /mcp suffix and Uri-based port in EmitBoundUrl; fix RestartRefusal
  pattern that never matched (is not var) and re-route observability command
  output from stderr back to stdout
- Fix proxy named-client mismatch so JsonRpcErrorHandler is actually wired
  (BackendSessionClient vs nameof(BackendSessions)); explicit BackendLauncher and
  ServerProbe factories (ambiguous ctors, missing TimeSpan budget)
- Migrate stale concrete SqliteConnectionFactory/SweepService/SharedExtractionRunner/
  WorkspaceService/SyncService ctor params to their interfaces
- Restore ArgumentNullException guards on WatchCommands/EncryptionCommands ctors

* test: migrate test and benchmark call sites to the refactored APIs

- SqliteMemoryStore new constructor order via TestData.CreateMemoryStore
- ConfigCommands/sub-commands RunAsync(StandardStreams) model; AppRunner-based
  CliCommandRunnerTests; indexer-safe OtlpNamesRegistry walk
- ServerProbe/BackendLauncher/ProxyForwarder/NodeRunner DI shapes; IdleWatchdog
  TimeSpan registration covered via the real host graph

* chore: record fix-after-refactor completion in state.json

* chore: record fix-after-refactor notes

+++ src/AiRaccoon.Core/Embedding/EmbeddingMath.cs
@@ -35,7 +35,7 @@ public static class EmbeddingMath
             return pooled;
         }
 
-        TensorPrimitives.Divide(pooled, (float)active, pooled);
+        TensorPrimitives.Divide(pooled, active, pooled);
 
         var norm = TensorPrimitives.Norm((ReadOnlySpan<float>)pooled);
         if (!(norm > 0))
